Разделил js код на модули.
