+++ index.html
@@ -8,7 +8,7 @@
   <title>Кексобукинг</title>
 </head>
 <body>
-  <script src="js/main.js" defer></script>
+  <script type="module" src="js/main.js" defer></script>
 
   <main>
     <div class="promo">
